[list] add list tests, fix some bugs

+++ data_structure/list/ft_list_size.c
@@ -12,9 +12,9 @@
 
 #include "ft_list.h"
 
-int	ft_list_size(t_list *begin_list)
+size_t	ft_list_size(t_list *begin_list)
 {
-	int		i;
+	size_t	i;
 
 	i = 0;
 	while (begin_list)
